Lead Generation/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
supabase==2.9.1
python-dotenv==1.0.1
groq==0.14.0
streamlit==1.49.0
pandas==2.2.3
requests==2.32.3
